+++ CHANGELOG.md
@@ -2,6 +2,20 @@
 
 All notable changes to the "brio" extension will be documented in this file.
 
-## Release 1.0.0
+### 1.1.0
+
+- Added the following commands:
+  - `Brio: Select Interpreter`
+  - `Brio: Clear Interpreter`
+  - `Brio: Start REPL`
+  - `Brio: Run Brio File in Terminal`
+
+- Added the ability to define a default and workspace Brio interpreter. 
+
+- Added a run/play icon for easier execution of Brio programs. 
+
+- Added improvements to syntax highlighting.
+
+## 1.0.0
 
 - Initial release with syntax highlighting support